refactor: extract direct node meta fast paths
